fix: compute router KL in fp32 to avoid small negative values in bf16 (#19)
At KL ~ 0 early in training, half-precision softmax/log-softmax rounding
produced kl_loss ~ -5e-4 on GH200. Divergence math now runs in fp32.

+++ src/exex/trainer.py
@@ -162,8 +162,10 @@ class ExpertTrainer:
                     ref["proj_weight"].to(device),
                 )
 
-            current_log_probs = F.log_softmax(current_logits, dim=-1)
-            ref_probs = F.softmax(ref_logits, dim=-1)
+            # fp32 for the divergence itself: at KL ~ 0 (early training),
+            # bf16/fp16 rounding yields small negative values
+            current_log_probs = F.log_softmax(current_logits.float(), dim=-1)
+            ref_probs = F.softmax(ref_logits.float(), dim=-1)
 
             kl = F.kl_div(current_log_probs, ref_probs, reduction="batchmean")
             total_kl = total_kl + kl
